+++ iam_child.tf
@@ -1,24 +1,28 @@
 #role in child account
 
 resource "aws_iam_role" "bastion_service_assume_role" {
-  name = "${var.environment_name}-${data.aws_region.current.name}-${var.vpc}_bastion"
+  name               = "${var.environment_name}-${data.aws_region.current.name}-${var.vpc}_bastion"
+  count              = "${local.assume_role_yes}"
+  assume_role_policy = "${data.aws_iam_policy_document.bastion_service_assume_role.json}"
+  tags               = "${var.tags}"
+}
 
+data "aws_iam_policy_document" "bastion_service_assume_role" {
   count = "${local.assume_role_yes}"
 
-  assume_role_policy = <<EOF
-{
-  "Version": "2012-10-17",
-  "Statement": [
-    {
-      "Action": "sts:AssumeRole",
-      "Principal": {
-        "Service": "ec2.amazonaws.com"
-      },
-      "Effect": "Allow"
+  statement {
+    actions = [
+      "sts:AssumeRole",
+    ]
+
+    principals {
+      type = "Service"
+
+      identifiers = [
+        "ec2.amazonaws.com",
+      ]
     }
-  ]
-}
-EOF
+  }
 }
 
 #Instance profile